+++ src/app/telefonos/telefonos.component.html
@@ -69,127 +69,11 @@
                
             </tr>
         </ng-template>
-        <ng-template pTemplate="body" let-telefonos let-i="rowIndex">
+        <ng-template pTemplate="body" let-telefonos >
             <tr> 
                 <td>
                    
-                    <p-accordion (onOpen)="getData(telefonos, i)" >
-                       
-                        <p-accordionTab header="{{telefonos.numero}}" >
-                            <div class="mb-3">
-                                <div class="mb-3 " >
-                                    
-                                <form [formGroup]="consumoForm" >
-                                    <p-card header="Crear nuevo consumo" >
-                                        
-                                        <div class="card-container"> 
-                                            <div class="inline-block mx-3">
-                                           
-                                             <p-inputNumber [showButtons]="true"  formControlName="consumo" mode="currency" currency="EUR" inputId="withoutgrouping" [useGrouping]="false"> </p-inputNumber>
-                                                  
-                                            </div>
-                                            <div class="inline-block mb-3 mx-3">
-                                                <span class="p-float-label">
-                                                    <p-calendar [showIcon]="true" view="month"   dateFormat="mm/yy" formControlName="fecha"  ></p-calendar>
-                                                    <label>Fecha</label>
-                                                </span>
-                                            </div>
-                                            <div class="block  mx-3 ">
-                                                <p-button label="Crear consumo"  (click)="crearConsumo(telefonos, i)" severity="success"></p-button>
-                                            </div>
-                                        </div>
-                                    </p-card>
-                                       
-                                   
-                                    
-                                </form>
-                                
-                            
-    
-                                </div>
-                                <div class="card-container">
-                                        <div class="inline-block  mx-3">
-                                            <p-toggleButton [(ngModel)]="checkedConsumo[i]" onLabel="Tabla" offLabel="Gráfica"></p-toggleButton>                                              
-                                       </div>
-
-                                       <div class="inline-block  mx-3">
-                                            <p-button label="Generar pdf" (click)="generatePDF(telefonos,i)" icon="pi pi-file-pdf"  ></p-button>
-                                       </div>
-
-                                       <div class="inline-block  mx-3">
-                                            <p-button label="Mandar por email" icon="pi pi-send" ></p-button>
-                                       </div>  
-                                
-                                </div>
-                              
-                            </div>
-                            
-                            
-                            
-                            <div *ngIf="emptychart[i]; else mostrarConsumos"> <!-- En caso de que no haya consumos-->
-                                <h3>No hay consumos apuntados</h3>
-                            </div>
-                        
-                        <ng-template #mostrarConsumos> <!-- Si hay consumos: -->
-
-                            <div *ngIf="!checkedConsumo[i]; else mostrarTabla"> 
-                                <!-- MOSTRAR GRÁFICAS-->
-                                <p-progressSpinner *ngIf="loading" [style]="{ width: '50px', height: '50px' }"></p-progressSpinner>
-                            
-
-                                <div style="display: block">
-                                    <p-chart type="line" id = "chart1{{i}}" [data]="chartData[i]" [options]="chartOptions"  ></p-chart>
-                                    <p-chart type="bar" [data]="chartDataPromedioMaxMin[i]" [options]="chartOptionsPromedioMaxMin" ></p-chart>
-                                </div>
-
-
-                            </div>
-                            
-                            <ng-template #mostrarTabla> 
-                                <!--MOSTRAR TABLA-->
-
-                            <div>
-                                <p-table [value]="consumos[i]" [tableStyle]="{ 'min-width': '50rem' }">
-                                    <ng-template pTemplate="header">
-                                        <tr>
-                                            <th>Consumo</th>
-                                            <th>Fecha</th>
-                                            <th>Opciones</th>
-                                            
-                                        </tr>
-                                    </ng-template>
-                                    <ng-template pTemplate="body" let-consumo>
-                                        <tr>
-                                            
-                                            <td>
-                                                <p-inputNumber [showButtons]="true"  [(ngModel)]="consumo.consumo" (keyup.enter)="editarConsumo(consumo, telefonos, i)"  mode="currency" currency="EUR" inputId="withoutgrouping" [useGrouping]="false"> </p-inputNumber>
-                                            </td>
-                                            <td>
-                                               
-                                                <span class="p-float-label">
-                                                    
-                                                    <p-calendar  appendTo="body"  [showIcon]="true" view="month" [showButtonBar]="true"  dateFormat="mm/yy"  [(ngModel)]="consumo.fecha" (onSelect)="editarConsumo(consumo, telefonos,i)"></p-calendar>
-                                                   
-                                                    <label> {{  consumo.fecha | date:'MM/yyyy' }}</label>
-                                                </span>
-                                            </td>
-                                            <td>                                           
-                                                <p-button label="Borrar Consumo"  (click)="eliminarConsumo(consumo.id, telefonos, i)" severity="danger" ></p-button>
-                                            </td>
-                                            
-                                        </tr>
-                                    </ng-template>
-                                </p-table>
-                            </div>
-
-
-
-                            </ng-template>
-                        </ng-template>       
-                      
-                            
-                        </p-accordionTab>
-                    </p-accordion>
+                 <app-consumo [telefono]="telefonos" > </app-consumo>
                 </td>
               
               <td> <p-button label="Borrar teléfono" (click)="borrarTelefono(telefonos.id)" severity="danger" ></p-button>
